refactor: HandlerAdapter에서 뷰를 보여주도록 변경
