+++ app/src/main/java/com/techcourse/DispatcherServlet.java
@@ -6,15 +6,11 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import java.util.LinkedList;
 import java.util.List;
-import java.util.Map;
 import java.util.Objects;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import webmvc.org.springframework.web.servlet.ModelAndView;
-import webmvc.org.springframework.web.servlet.View;
 import webmvc.org.springframework.web.servlet.mvc.tobe.HandlerAdapter;
 import webmvc.org.springframework.web.servlet.mvc.tobe.HandlerMapping;
-import webmvc.org.springframework.web.servlet.view.JspView;
 
 public class DispatcherServlet extends HttpServlet {
 
@@ -60,30 +56,10 @@ public class DispatcherServlet extends HttpServlet {
                     .findAny()
                     .orElseThrow();
 
-            final Object result = handlerAdapter.execute(request, response, handler);
-
-            if (result instanceof String) {
-                move((String) result, request, response);
-                return;
-            }
-            final ModelAndView mav = (ModelAndView) result;
-            final Map<String, Object> model = mav.getModel();
-            final View view = mav.getView();
-            view.render(model, request, response);
+            handlerAdapter.execute(request, response, handler);
         } catch (Throwable e) {
             log.error("Exception : {}", e.getMessage(), e);
             throw new ServletException(e.getMessage());
         }
-    }
-
-    private void move(final String viewName, final HttpServletRequest request, final HttpServletResponse response)
-            throws Exception {
-        if (viewName.startsWith(JspView.REDIRECT_PREFIX)) {
-            response.sendRedirect(viewName.substring(JspView.REDIRECT_PREFIX.length()));
-            return;
-        }
-
-        final var requestDispatcher = request.getRequestDispatcher(viewName);
-        requestDispatcher.forward(request, response);
     }
 }
