+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/asis/ManualHandlerAdapter.java
@@ -3,6 +3,7 @@ package webmvc.org.springframework.web.servlet.mvc.asis;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import webmvc.org.springframework.web.servlet.mvc.tobe.HandlerAdapter;
+import webmvc.org.springframework.web.servlet.view.JspView;
 
 public class ManualHandlerAdapter implements HandlerAdapter {
 
@@ -12,9 +13,16 @@ public class ManualHandlerAdapter implements HandlerAdapter {
     }
 
     @Override
-    public Object execute(final HttpServletRequest request, final HttpServletResponse response, final Object handler)
+    public void execute(final HttpServletRequest request, final HttpServletResponse response, final Object handler)
             throws Exception {
         final Controller controller = (Controller) handler;
-        return controller.execute(request, response);
+        final String viewName = controller.execute(request, response);
+        if (viewName.startsWith(JspView.REDIRECT_PREFIX)) {
+            response.sendRedirect(viewName.substring(JspView.REDIRECT_PREFIX.length()));
+            return;
+        }
+
+        final var requestDispatcher = request.getRequestDispatcher(viewName);
+        requestDispatcher.forward(request, response);
     }
 }
